+++ src/mainwindow_p.h
@@ -12,6 +12,8 @@
 
 #include "mainwindow.h"
 
+#include "MViewerVersion.h" // M24 version SSOT (generated from CMake project VERSION)
+
 #include "application/OpenDirectoryUseCase.h"
 #include "appstate.h"
 #include "core/EventBus.h"